adding Wrapper file with convenient type classes
